Move FastAPI source to app package and adjust Docker
